Resolved conflicts with moving of code
